Smal refinements
- display current distance readings for the right sensor
- voltagemeter tweaks, use ESP lib rather arduino
- some typos removed (and likely added)
- reduced number of decimal digits in CSV where we do not have more
precision any way

+++ OpenBikeSensorFirmware/displays.h
@@ -198,7 +198,7 @@ class SSD1306DisplayDevice : public DisplayDevice
 
     void showValues(
       HCSR04SensorInfo sensor1, HCSR04SensorInfo sensor2,
-      int minDistanceToConfirm, int lastMesurements, boolean insidePrivacyArea);
+      uint16_t minDistanceToConfirm, int lastMesurements, boolean insidePrivacyArea);
 
 };
